Basic scaffold for Rolodex project
This has slightly more than the empty application scaffold, including a
Contact model and a Rolodex collection (of Contacts).

Neither of those contain any functional code, but there is a (failing)
test that expects the Contact model to have a name attribute. This will
hopefully serve as an example of how to write tests for model classes.